Tynset: defect fix.
Found a race where a vbucket was quickly created and deleted and the flusher managed to try and flush it
(because of checkpoint.cc triggers flusher for meta keys i.e. cp_start/end).

Even though the VB doesn't exist (you can't get a VB pointer) the bucketCreation flag is still true

Fix is to keep consistency amongst the vbmap entries and the creation array

Change-Id: Id85629f65763827a5551ba0d79720e21189df52c