Tynset: Defect fix (push upstream)
Found that a sequence of key operations allows the flusher to miss a flush.

If the flusher sees

set(key1)
del(key)

and prev == key1

Then the del will be ignored.

This happens if the flusher runs between checkpoints and the operations are in different checkpoints.
Within a single checkpoint the set/del are collapsed so only the del happens.

Change-Id: I36e9056c397638f0a63ebd5ec25939b34d9ef2fa