Tynset: HashTable storage split.
The underlying storage (values array) and mutexes of the HashTable are now moved to their
own class (HashTableStorage).

HashTableStorage is the object that does not get duplicated when new buckets/vbuckets are created.
HashTable is an object that is created for every vbucket, this is to allow correct stats tracking.

This patch also fixes cleanup paths so that destructing objects correctly deletes the correct StoredValues.
You can no longer clear a HashTable and it affect every bucket that uses the underlying HashTableStorage.

PauseResumeVisiting is also fixed in this patch so that the defragmenter now runs as a StoragePool
task rather than 1 per bucket

Change-Id: I431d5f8deb484d85a32e255afdda48f5ab25dcc8