ItemKey and StoredValueKey
To bring control of how item's are identified into ep-engine this patch replaces
use of std::string for identifying items. It is not saying std::string must be,
just that when items are destined for storage, if we use the new ItemKey class
we can identify keys in new ways.

ItemKey replaces the use of std::string throughout many parts of ep-engine and
allows the code to carry around a little more information than just key/keylen.
We add the bucket_id_t to the type which will in the future to allow the code
to use shared hashtables to store items, allowing bucket1:key and bucket2:key to
be correctly differentiated. ItemKey is generally used on the IO threads and
destructs once the thread is complete and the key has been moved to a hashtable
or just not needed.

The ItemKey implementation basically takes the incoming key and glues a bucket-id
onto the bytes enabling the differentiation of same key for different buckets.

A second key type is introduced for the actual in-memory storage (when ItemKey needs
to move to the HashTable). StoredValueKey is similar in concept, but ensures works
in conjunction with StoredValue to use the continous memory that StoredValueFactory
creates for the key and value.

The bucket-id is not being stored (not yet accounted in StoredValue).

In general many parts of the code now have the following change, and interface changes
to move std:string to ItemKey.

-std::string key(packet->key, packet->len);
+ItemKey key(packet->key, packet->len, engine.getBucketId());

100% pass on ep-engine make test.

Change-Id: I6495bb6348466eae346b571ef039acda68247e2e