Experimental TCP flow-control
Create a "mode" config flag, this determines the action
a consumer takes to manage flow.

a) dcp-managed - i.e. we will ack bytes using DCP_ACKNOWLEDGE
b) ewouldblock - we manage flow by ewouldblock and pausing, tcp
   will buffer if need be and trigger a pause on producer.

The existing policy flag is used to calc the consumer buffer size
and that size is used for both modes.

Note that policy - none sets the buffer to max(uint64) so that flow
control never triggers.

Change-Id: I01faa55efcbe97c108f684b7d5230da0137f1cfb