MB-18452: Single threaded test harness improvements
Refactor parts of the very new evp_store_single_threaded_test so that
it's simpler to drive tasks making new tests easier to write.

The main change is to provide helper methods for running any task from
a queue (with some checks) and a way to push a clean shutdown.

Change-Id: I19782dcacb36048151bc073a377f28603ff83033