MB-16181: Basic support for meta-events
This adds the ability to queue into a checkpoint
a meta-event. This records in the data stream
at the current seqno an event has occured.

This commit reads the events back, but hides them from
DCP output (they just don't appear to DCP streams).

Future commits will add support to DCP to allow the events to
form real packets for network transmission.

The value of the events is currently undefined and risks
weirdness if a master->future-master upgrade occurs with
events on disk.

This commit does not actually enable writing of events from
any public API.

Change-Id: I2ac3cf6b275f4bfb8e3ecda086174b15ad94ba82