MB-19948: Refactor CouchKVStore meta-data handling
Create classes to handle the KVStore metadata rather
than working on raw buffers with memcpy and offsets.

This includes a fix for the orginal MB-19948 issue in that
we correctly read V1 metadata (the 18byte version written
by 3.x). The fix is kind of hidden in the refactor, but the
new code explictly looks at the size of metadata and has
construction paths for 16, 18 and 19 bytes.

Change-Id: I7d2042af519819737a6e4d070fd8185c84778d6a