MB-16181: Reduce printing of keys
This commit removes a number of places where the key was being
printed (to logs).

This commit also cleans some logging so that messages can be tied to
code quicker (function names logged).

Where possible, a seqno/vb is printed instead of the raw key.

If the key must be printed, it is wrapped in braces to allow easy
removal of keys from logs should it be required.

Change-Id: Ieb191d0d9445b3fa7d9a1cb495aee603e80b29c0