MB-16181: std::string to our own StorageKey class
We use std::string to represent document keys throughout. Collections
will require that a document's key can be interogated in different
and we require our own container class for a key.

Introducing StorageKey

This commit introduces a replacement for std::string called StorageKey
which provides a minimal interface, but some important methods to
build on. A StorageKey is how you would locate a document within
ep-engine's storage mediums, primarily the HashTable and persisted
data stores but also transient storage like checkpoint queues.

Introducing ProtocolKey

* StorageKey::getProtocolKey()

A protocol key is the part of the key owned by the external protocol,
for example the user's keys in a memcached binary SET command.
A protocol key can always be obtained from the StorageKey, but a
StorageKey cannot be infered from a ProtocolKey (without extra
information).

In this introductory commit, the ProtocolKey.data()/.size() is
logically the same as the StorageKey.data()/.size(), but that will be
changing.

Collections and ProtocolKey/StorageKey

The goal is that the introduction of collections will mean that a
client may place a "a::b" document into many of a collection or
default storage, and internally ep-engine will represent each of these
documents via StorageKey's that yield different data()/size() values
(and thus don't collide in hashtables/b-trees), yet these different
StorageKeys will yield the same protocol key value of "a::b"

Change-Id: I50ffd064c2ecee715d554531b331fc869df09679