Reinstate 685c96d2
History:
1. 685c96d2 partially reverted 769cfe0 to remove the now defunct
   commit-interval code.
2. 685c96d2 was thought to trigger a simple-test failure so 78dff039
   reverted 685c96d2
3. The simple-test failure look incorrect so this commit re-instates
   685c96d2

I.e. this is now a revert of a revert...

The 685c96d2 commit message was (which now applies to this commit)

<begin 685c96d2>
Partially revert 769cfe0
Note this is a heavily fixed "git revert" due to conflicts...

Remove parts of 769cfe0 (MB-14190) which are no longer needed, that is
the commit interval code and the unnecessary extra calls to
store->commit.

The orginal commit added the interval code to support ForestDB
which needed to batch file commits based on some configuration value,
whereas CouchStore commits per vbucket update (i.e. always had an
interval of 1).

The orginal commit also then added extra store->commit calls which in
the couchstore configuration meant calling down a few layers only to
return once the CouchKVStore detected all items were flushed. This is
wasted cycles for our supported configuration.

Note that some of the refactoring created by 769cfe0 remains.
<end 685c96d2>

Change-Id: I3ca6cf00a3c0cd4c3dabe9d4342e463a231a45cd