MB-16181: Add more ostream operators and dump methods
Add more ostream operators to the Collections classes and also give
them dump() methods so we can easily trigger output to stderr.

Finally add a hook so we can force the logging of a lot of collection
state, this will eventually be migrated to stats and more supported
debug methods.

Change-Id: I18abe41cd8f00fe18cb4fb972fcc054375f02d39